refacto: check::mess by adding namespace mess_
