+++ src/builder/command-validation/checkers/check.cpp
@@ -30,7 +30,7 @@ bool check::target(CmdSpec &cmd, int idx) {
 	return (true);
 }
 
-bool targetIsOnChan(CmdSpec &cmd, int idx) {
+bool check::targetIsOnChan(CmdSpec &cmd, int idx) {
 	(void)cmd;
 	(void)idx;
 	return (true);
@@ -39,44 +39,6 @@ bool targetIsOnChan(CmdSpec &cmd, int idx) {
 bool check::invite(CmdSpec &cmd, int idx) {
 	(void)cmd;
 	(void)idx;
-	return (true);
-}
-
-bool check::mess(CmdSpec &cmd, int idx) {
-	(void)idx;
-	if (cmd[target_].empty()) {
-		reply::send_(
-			cmd.getSender().getFd(),
-			ERR_NORECIPIENT(cmd.getSender().cliInfo.getNick(), cmd.getName()));
-		return (false);
-	}
-	if (cmd[message_].empty()) {
-		reply::send_(cmd.getSender().getFd(),
-					 ERR_NOTEXTTOSEND(cmd.getSender().cliInfo.getNick()));
-		return (false);
-	}
-	size_t i = 0;
-	while (i < cmd[target_].size()) {
-		if (cmd[target_][i][0] != '#' ||
-			(cmd[target_][i][0] != '@' &&
-			 (cmd[target_][i].size() > 1 && cmd[target_][i][1] != '#'))) {
-			if (!check::target(cmd, i)) {
-				cmd[channel_].rmParam(i);
-				continue;
-			}
-		} else {
-			if (!check::chan(cmd, i) && !check::chans_::isOnChan(cmd, i)) {
-				cmd[channel_].rmParam(i);
-				continue;
-			}
-			if (cmd[target_][i][0] == '@') {
-				//TODO: add bool only op et rm le @ s'il est la
-			}
-		}
-		i++;
-	}
-	if (!cmd[channel_].size())
-		return (false);
 	return (true);
 }
