+++ src/builder/command-validation/checkers/check_mess.cpp
@@ -1,0 +1,53 @@
+//HEADER
+
+#include "validator.hpp"
+
+bool check::mess(CmdSpec &cmd, int idx) {
+	(void)idx;
+	if (!check::mess_::params(cmd))
+		return (false);
+
+	size_t i = 0;
+	while (i < cmd[target_].size()) {
+		if (check::mess_::isNick(cmd[target_][i])) {
+			if (!check::target(cmd, i)) {
+				cmd[channel_].rmParam(i);
+				continue;
+			}
+		} else {
+			if (!check::chan(cmd, i) || !check::chans_::isOnChan(cmd, i)) {
+				cmd[channel_].rmParam(i);
+				continue;
+			}
+			if (cmd[target_][i][0] == '@') {
+				//TODO: add bool only op et rm le @
+			}
+		}
+		i++;
+	}
+	if (!cmd[channel_].size())
+		return (false);
+	return (true);
+}
+
+bool check::mess_::params(CmdSpec &cmd) {
+	if (cmd[target_].empty()) {
+		reply::send_(
+			cmd.getSender().getFd(),
+			ERR_NORECIPIENT(cmd.getSender().cliInfo.getNick(), cmd.getName()));
+		return (false);
+	}
+	if (cmd[message_].empty()) {
+		reply::send_(cmd.getSender().getFd(),
+					 ERR_NOTEXTTOSEND(cmd.getSender().cliInfo.getNick()));
+		return (false);
+	}
+	return (true);
+}
+
+bool check::mess_::isNick(std::string &target) {
+	if (target[0] != '#' ||
+		(target[0] != '@' && (target.size() > 1 && target[1] != '#')))
+		return (true);
+	return (false);
+}
